requirements.txt
PyYAML>=6.0
python-dotenv>=1.0.1
requests>=2.32.3
pandas>=2.2.2
absl-py>=2.2.0
langdetect>=1.0.9
immutabledict>=4.2.1
nltk>=3.9.1
pyarrow>=16.1.0
torch>=2.0.0
transformers>=4.40.0
